pyproject.toml
[project]
name = "prompt-interp"
version = "0.1.0"
description = "SONAR-LLM experiments - sentence-level language modeling"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.8.0",
    "sonar-space>=0.5.0",
    "nltk>=3.8",
    "transformers>=4.40.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "datasets>=4.4.2",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "fairseq2"
url = "https://fair.pkg.atmeta.com/fairseq2/whl/pt2.8.0/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
fairseq2 = { index = "fairseq2" }
